ams-core:  Fix GNE, add ReadDateTime / WaitMouseUp
Fix GetNextEvent() to filter activate and update events according to the
event mask.

Implement ReadDateTime() as an accessor for Time.  It's supposed to read
the clock chip, but AMS /has/ no clock chip.  AMS needs no clock chip --
the low memory global variable Time always has the right value in it.  A
normal, well-behaved Mac application shouldn't be calling ReadDateTime()
in the first place, but some applications can be naughty and mischievous
little programs, oh yes.  But it's far too late to complain about it now
when an application was developed in 1984 and has become enmeshed in the
history we seek to preserve.

Implement WaitMouseUp().

This is a debug build, as of 90447e96b922d83e996e16b560d4204d243c5c43 in
metamage_1.git.