ams-io:  Make breaking changes to sound protocol
The big one is that the length field's value now excludes the mode field
in its measurement, so you'll need to add two (or cease subtracting two)
to get the same value as before.  Also important is that the KillIO call
now sends 0x5858 ('XX'), not 0xFFFF.

This is a debug build, as of 73965bbeeb55bf36043881d4d01d0677465d196a in
metamage_1.git.