ams-ui:  Add more text editing building blocks
It's not even close to done yet, but this reduces unimplemented-trap and
related crashes (while leaving text undrawn for now).  Also, we've added
StringToNum() and TestControl(), plus dialog TextEdit records as needed.

This is a debug build, as of 031a2b552ffa49d17a90da3bcc0224f7778b0e19 in
metamage_1.git.