ams-qd:  Secure UnpackBits(), actually PackBits()
Reject $80 metadata bytes in packed data, and truncate runs so as not to
overrun the specified size of the output buffer.  Actually pack bytes in
PackBits() (which could be relevant if output is assumed to be smaller).

This is a debug build, as of 2417174ac05881a40c7c7088eaf372a965345247 in
metamage_1.git.