Lode Runner:  Opt into auto-vsync
I'm not seeing the tearing I used to see in Amethyst with either that or
Amphitheatre without auto-vsync now, but flickering is noticeable in the
brick-burning animation, so we'll want to keep auto-vsync on.  To do so,
we need to explicitly request it before it stops being the default mode.