ams-core:  Force-throttle GNE after 3 null events
This prevents Lode Runner from pegging a core while it's idle and not in
a modal dialog (e.g. at the splash screen, when the player avatar blinks
before the start of a level, and when the user enters a new high score).

This is a debug build, as of a568636142e178e8ce07be718c7699768d308886 in
metamage_1.git.