ams-qd:  Save/restore `line` pnLoc; fix CopyMask()
Restore the previous pen location after processing `line` opcodes.  This
fixes the Info screens of Dark Castle and System's Twilight's About/Help
screens (when sufficient depth is present to draw lines in low contrast).

Also, don't pass garbage to blit_bytes() in the high words of D0 and D1.
This will allow for passing additional blitting arguments in the future.

This is a debug build, as of 7b8e043be6eb2a52e82840596800879ef722a020 in
metamage_1.git.